+++ QBCore_2F0666.base/resources/[qb]/qb-smallresources/client/point.lua
@@ -55,4 +55,4 @@ RegisterCommand('point', function()
     end
 end, false)
 
-RegisterKeyMapping('point', 'Toggles Point', 'keyboard', 'B')
+RegisterKeyMapping('point', 'İşaret Et', 'keyboard', 'B')